test: verify game floating point scope and thread isolation

+++ include/kinoko/game_math.h
@@ -7,7 +7,7 @@ extern "C" {
    Set both x87 and SSE so migrated C++ follows the original x87 policy. */
 uint32_t kinoko_enter_game_math(void);
 void kinoko_leave_game_math(uint32_t previous_rounding);
-uint32_t kinoko_run_game_math(uint32_t (__stdcall *update)(void *), void *argument);
+unsigned long kinoko_run_game_math(unsigned long (__stdcall *update)(void *), void *argument);
 #ifdef __cplusplus
 }
 #endif
